+++ crates/assistant/src/prompts.rs
@@ -1,3 +1,3 @@
-pub mod custom_prompts;
 pub mod prompt_library;
+pub mod prompt_manager;
 pub mod prompts;
